requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
celery>=5.3
redis>=5.0
openai>=1.30
gunicorn>=21.2
